services/function-preparation/requirements.txt
asyncpg
kafka-python
